+++ Graph.h
@@ -11,7 +11,7 @@ private:
 public:
     Graph(int n);
     ~Graph();
-    void set_dis(int i, int j, int dis); // Set the distance between i and j
+    void set_dis(int i, int j, float dis); // Set the distance between i and j
     float get_dis(int i, int j);
     int get_city_num();
 };
